added __init__ to Product_insights/Twitter

+++ Product_Insights/Twitter/create_twitter_tables.py
@@ -9,7 +9,7 @@ def create_twitter_sentiment():
       bigquery.SchemaField('created_at', 'TIMESTAMP', mode='NULLABLE'),
       bigquery.SchemaField('full_text', 'STRING', mode='NULLABLE'),
       bigquery.SchemaField('user_id', 'INTEGER', mode='NULLABLE'),
-      bigquery.SchemaField('in_reply_to_status_id_str', 'INTEGER', mode='NULLABLE'),
+      bigquery.SchemaField('in_reply_to_status_id_str', 'FLOAT', mode='NULLABLE'),
       bigquery.SchemaField('score', 'FLOAT', mode='NULLABLE',
                            description='Sentiment score from the google language api'),
       bigquery.SchemaField('magnitude', 'FLOAT', mode='NULLABLE',
